Rewrite match_instr! as a proc macro
The declarative macro started out simple, but quickly became quite
complicated.

There are a few advantages to using a proc macro:
1. It can generate significantly less code. Rather than recursively
   expanding to nested `match` statements, a simple state machine is
   generated that handles matching the user-provided patterns and
   extracting their bound values, with a second match binding the values
   and executing the user-provided match arm bodies. This drops the
   amount of code generated for a `match_instr!` with 6 cases from ~600
   lines to ~75, and it now grows linearly with additional cases rather
   than exponentially.
2. It is much easier to support additional functionality. In the process
   of rewriting this I added support for `@` bindings to both
   instructions and expressions, and for `|` patterns.
3. Better error messages.
4. It's mostly normal Rust code so hopefully it'll be easier to
   come back to in the future.

The big disadvantage is that it's a lot more code.

While I was here I also added a bunch of test cases to make sure what
used to work still works, and to ensure things did not regress as I
refactored it.