Improve how match_instr! binds variables
It is now possible to bind directly to the `LowLevelILExpression` of
both expressions and sub-expressions. This required reworking how
variables in the pattern are tracked so that both simple and complex
patterns bind variables in the same way, and so that both top-level and
sub-expressions use the same code path for matching patterns and binding
variables.